% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoanalysis.R
\name{autoanalysis}
\alias{autoanalysis}
\title{autoanalysis}
\usage{
autoanalysis(mc = TRUE, env = NULL)
}
\arguments{
\item{mc}{logical enables or disables the use of multiple cores}
}
\value{
spikes processed spike data
}
\description{
This is the starting point of the analysis. The function will analyse ephys data with default parameters. The returned object contains all the data from the data directory. This object can then be handed over to the batch_spikes function for further analysis.
}

