% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_png_overview.r
\name{create_png_overview}
\alias{create_png_overview}
\title{\code{create_png_overview} A function to create png overviews of ephys log files}
\usage{
create_png_overview(wd = NA)
}
\arguments{
\item{wd}{working directory where the data is. It is required to use the folder containing the data, meta, and png subfolders as input}
}
\description{
\code{create_png_overview} A function to create png overviews of ephys log files
}

