% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.gauss.R
\name{filter.gauss}
\alias{filter.gauss}
\title{filter.gauss}
\usage{
\method{filter}{gauss}(signal, sigma)
}
\arguments{
\item{signal}{}

\item{sigma}{}
}
\description{
\code{filter.gauss} this function applies a gaussian convolution filter to the input signal. the window width is specified by sigma
}

