% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gauss_filter.r
\name{gauss_filter}
\alias{gauss_filter}
\title{\code{gauss_filter} this function applies a gaussian convolution filter to the input signal. the window width is specified by sigma}
\usage{
gauss_filter(signal, sigma)
}
\arguments{
\item{signal}{}

\item{sigma}{}
}
\description{
\code{gauss_filter} this function applies a gaussian convolution filter to the input signal. the window width is specified by sigma
}

