% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.H1signal.R
\name{get.H1signal}
\alias{get.H1signal}
\title{get_H1signal}
\usage{
get.H1signal(data)
}
\arguments{
\item{data}{data from one recording file (2 channels)}
}
\value{
the data in a data frame
}
\description{
This file is part of the rephys package.
}
\details{
Copyright(c) Martin Rüßler

This file may be licensed under the terms of of the
GNU General Public License Version 3 (the ``GPL''),
or (at your option) any later version.

Software distributed under the License is distributed
on an ``AS IS'' basis, WITHOUT WARRANTY OF ANY KIND, either
express or implied. See the GPL for the specific language
governing rights and limitations.

You should have received a copy of the GPL along with this
program. If not, go to http://www.gnu.org/licenses/gpl.html
or write to the Free Software Foundation, Inc.,
51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.

The development of this software was supported by the
Excellence Cluster EXC 277 Cognitive Interaction Technology.
The Excellence Cluster EXC 277 is a grant of the Deutsche
Forschungsgemeinschaft (DFG) in the context of the German
Excellence Initiative.

\code{get_H1signal} Since the data are not packed in a container together with metadata it is a simplified version.
}

