% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.isipsth.R
\name{get.isipsth}
\alias{get.isipsth}
\title{get.isipsth}
\usage{
get.isipsth(spikes, binwidth = 10, samplerate = 25000)
}
\arguments{
\item{spikes}{matrix of spike data}

\item{binwidth}{in milliseconds}

\item{samplerate}{in Hertz}
}
\value{
isipsth matrix
}
\description{
calculate an inter-spike-interval histogram
}

