% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.psth.R
\name{get.psth}
\alias{get.psth}
\title{get.psth}
\usage{
get.psth(spikes, binwidth = 10, samplerate = 25000)
}
\arguments{
\item{spikes}{a matrix of spike data}

\item{binwidth}{im ms}

\item{samplerate}{in Hertz}
}
\value{
psth data
}
\description{
calculate a PSTH histogram
}

