% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_H1signal.r
\name{get_H1signal}
\alias{get_H1signal}
\title{\code{get_H1signal} a copy of the matlab function with the same name. Since the data are not packed in a container together with metadata it is a simplified version.}
\usage{
get_H1signal(data)
}
\arguments{
\item{data}{data from one recording file (2 channels)}
}
\value{
the data in a data frame
}
\description{
\code{get_H1signal} a copy of the matlab function with the same name. Since the data are not packed in a container together with metadata it is a simplified version.
}

