% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.data.R
\name{plot.data}
\alias{plot.data}
\title{plot.data}
\usage{
\method{plot}{data}(spikes, filelist, stimlist, outputdir)
}
\arguments{
\item{spikes}{data frame of spike data}

\item{filelist}{}

\item{stimlist}{}

\item{outputdir}{}
}
\value{
nothing yet
}
\description{
plot.data
}

