% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.isipsth.R
\name{plot.isipsth}
\alias{plot.isipsth}
\title{plot.isipsth}
\usage{
\method{plot}{isipsth}(spikes, binwidth, samplerate)
}
\arguments{
\item{spikes}{a spike data frame}

\item{binwidth}{width of the bin}

\item{samplerate}{the samplerate of the data}
}
\value{
nothing yet
}
\description{
plot.isipsth
}

