% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.png.overview.R
\name{plot.png.overview}
\alias{plot.png.overview}
\title{plot.png.overview}
\usage{
\method{plot}{png.overview}(wd = NA, maxchunk = 100)
}
\arguments{
\item{wd}{character working directory where the data is. It is required to use the folder containing the data, meta, and png subfolders as input}

\item{maxchunk}{numeric add a maximum number of files to be processed at once}
}
\description{
\code{plot.png.overview} A function to create png overviews of ephys log files
}
\details{
This file is part of the rephys package.

Copyright(c) Martin Rüßler

This file may be licensed under the terms of of the
GNU General Public License Version 3 (the ``GPL''),
or (at your option) any later version.

Software distributed under the License is distributed
on an ``AS IS'' basis, WITHOUT WARRANTY OF ANY KIND, either
express or implied. See the GPL for the specific language
governing rights and limitations.

You should have received a copy of the GPL along with this
program. If not, go to http://www.gnu.org/licenses/gpl.html
or write to the Free Software Foundation, Inc.,
51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.

The development of this software was supported by the
Excellence Cluster EXC 277 Cognitive Interaction Technology.
The Excellence Cluster EXC 277 is a grant of the Deutsche
Forschungsgemeinschaft (DFG) in the context of the German
Excellence Initiative.
}

