% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.quickmatrix.R
\name{plot.quickmatrix}
\alias{plot.quickmatrix}
\title{plot.quickmatrix}
\usage{
\method{plot}{quickmatrix}(matrix, filename = "spike")
}
\arguments{
\item{matrix}{is the data matrix}

\item{filename}{is the filename prefix}
}
\description{
\code{plot.quickmatrix} This function plots each column of a data matrix
}

