% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.raster.R
\name{plot.raster}
\alias{plot.raster}
\title{plot.raster}
\usage{
\method{plot}{raster}(spikes, sr)
}
\arguments{
\item{spikes}{}

\item{sr}{sampling rate}
}
\description{
Create a raster plot with all the spike data
}

