% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.spikes.R
\name{plot.spikes}
\alias{plot.spikes}
\title{plot.spikes}
\usage{
\method{plot}{spikes}(spikes, filelist, stimlist, outputdir)
}
\arguments{
\item{spikes}{}

\item{filelist}{}

\item{stimlist}{}

\item{outputdir}{}
}
\value{
nothing
}
\description{
plot.spikes
}

