% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_quickmatrix.r
\name{plot_quickmatrix}
\alias{plot_quickmatrix}
\title{\code{startanalysis} This function plots each column of a data matrix}
\usage{
plot_quickmatrix(matrix, filename = "spike")
}
\arguments{
\item{matrix}{is the data matrix}

\item{filename}{is the filename prefix}
}
\description{
\code{startanalysis} This function plots each column of a data matrix
}

