% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotstimulus.r
\name{plotstimulus}
\alias{plotstimulus}
\title{plotstimulus}
\usage{
plotstimulus(stimdf)
}
\arguments{
\item{stimdf}{a data frame containing stimulus data}
}
\description{
add stimuli to a plot. the stimuli data contain information about the position of the stimuli
}

