% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.spikes.R
\name{prune.spikes}
\alias{prune.spikes}
\title{prune.spikes}
\usage{
prune.spikes(spikes, min.isi)
}
\arguments{
\item{spikes}{spike matrix to be pruned}

\item{min.isi}{minimum inter stimulus interval in milliseconds}
}
\value{
a matrix containing a pruned spike
}
\description{
a copy of the H1toolbox prune_spikes function. This Function removes spikes that fall within the minimum inter stimulus interval of the previous spike. It is safe to assume that a spike within 3 ms can be filtered out.
}

