% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prunespikes}
\alias{prunespikes}
\title{prunespikes}
\arguments{
\item{spikes}{spike matrix to be pruned}

\item{minisi}{minimum inter stimulus interval in milliseconds}
}
\value{
a matrix containing a pruned spike
}
\description{
A C++ implementation of a demanding function
}

