% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.ephysdata.R
\name{read.ephysdata}
\alias{read.ephysdata}
\title{read.ephysdata}
\usage{
read.ephysdata(files, folder, mc = TRUE)
}
\arguments{
\item{files}{a list of filenames}

\item{folder}{character a folder path to read ephys data from}

\item{mc}{enables or disables the use of multiple cores}
}
\value{
a data frame with ephys data
}
\description{
a function to read ephys log files.
}

