% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.metadata.R
\name{read.metadata}
\alias{read.metadata}
\title{read.metadata}
\usage{
read.metadata(datafolder, metafolder)
}
\arguments{
\item{datafolder}{where the data files are located}

\item{metafolder}{where the metadata files are located}
}
\value{
returns a list with metadata for all files from the data folder
}
\description{
a function to read ephys metadata files.
}

