% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.stimdata.R
\name{read.stimdata}
\alias{read.stimdata}
\title{read.stimdata}
\usage{
read.stimdata(metalist,
  stimlibrary = "/home/martin/datarepos/ephysstimlibrary/")
}
\arguments{
\item{metalist}{list of metadata data frames}

\item{stimlibrary}{character folder where the stim library is}
}
\value{
a list with stimulus data frames
}
\description{
match the experiment parameters with a filename in the stimlibrary. the function takes the output of the readmetadata function as an input. from that data it constructs the file paths to the stimulus files. example filename from the library: ephyslog-amp25-avgpos100-freq0.5.log
}
\details{
This file is part of the rephys package.

Copyright(c) Martin Rüßler

This file may be licensed under the terms of of the
GNU General Public License Version 3 (the ``GPL''),
or (at your option) any later version.

Software distributed under the License is distributed
on an ``AS IS'' basis, WITHOUT WARRANTY OF ANY KIND, either
express or implied. See the GPL for the specific language
governing rights and limitations.

You should have received a copy of the GPL along with this
program. If not, go to http://www.gnu.org/licenses/gpl.html
or write to the Free Software Foundation, Inc.,
51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.

The development of this software was supported by the
Excellence Cluster EXC 277 Cognitive Interaction Technology.
The Excellence Cluster EXC 277 is a grant of the Deutsche
Forschungsgemeinschaft (DFG) in the context of the German
Excellence Initiative.
}
\seealso{
readmetadata
}

