% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readephysdata.r
\name{readephysdata}
\alias{readephysdata}
\title{readephysdata}
\usage{
readephysdata(files, folder, mc = TRUE)
}
\arguments{
\item{files}{a list of file names}

\item{folder}{character a folder path to read ephys data from}

\item{mc}{enables or disables the use of multiple cores}
}
\value{
a data frame with ephys data
}
\description{
a function to read ephys log files.
}

