% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readstimdata.r
\name{readstimdata}
\alias{readstimdata}
\title{readstimdata}
\usage{
readstimdata(metalist,
  stimlibrary = "/home/martin/datarepos/ephysstimlibrary/")
}
\arguments{
\item{metalist}{list of metadata data frames}

\item{stimlibrary}{character folder where the stim library is}
}
\value{
a list with stimulus data frames
}
\description{
match the experiment parameters with a filename in the stimlibrary. the function takes the output of the readmetadata function as an input. from that data it constructs the file paths to the stimulus files. example filename from the library: ephyslog-amp25-avgpos100-freq0.5.log
}
\seealso{
readmetadata
}

