% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/startanalysis.r
\name{startanalysis}
\alias{startanalysis}
\title{startanalysis}
\usage{
startanalysis(mc = TRUE)
}
\arguments{
\item{mc}{logical enables or disables the use of multiple cores}
}
\value{
ephys raw data to be further processed
}
\description{
This is the starting point of the analysis. The function sources some requisites, and sets the working directory (Details of this will change). The returned object contains all the data from the data directory. This object can then be handed over to the batch_spikes function for further analysis.
}

