% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threshold.spikes.R
\name{threshold.spikes}
\alias{threshold.spikes}
\title{threshold.spikes}
\usage{
threshold.spikes(signal, threshold)
}
\arguments{
\item{signal}{numeric vector of ephys data}

\item{threshold}{numeric value}
}
\value{
a matrix with spikes
}
\description{
\code{threshold.spikes} apply a threshold to the input signal and return a vector of 0 and 1
}

