% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threshold_spikes.r
\name{threshold_spikes}
\alias{threshold_spikes}
\title{\code{threshold_spikes} apply a threshold to the input signal and return a vector of 0 and 1}
\usage{
threshold_spikes(signal, threshold)
}
\arguments{
\item{signal}{numeric vector of ephys data}

\item{threshold}{numeric value}
}
\value{
a matrix with spikes
}
\description{
\code{threshold_spikes} apply a threshold to the input signal and return a vector of 0 and 1
}

