% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{timer}
\usage{
timer(time.start = FALSE)
}
\arguments{
\item{time.start}{a time object}
}
\value{
eiter a time or a time difference depending on how the function was called
}
\description{
can be called both with and without a time point
}
\examples{
function.start <- timer()
Sys.sleep(3)
time.diff <- timer(function.start)
print(paste("Function took", time.diff, "seconds."))
}

