% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vplayout.R
\name{vplayout}
\alias{vplayout}
\title{vplayout}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{x coordinate of the viewport}

\item{y}{y coordinate of the viewport}
}
\value{
viewport layout
}
\description{
a small convenience function
}

