% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.range.R
\name{zero.range}
\alias{zero.range}
\title{zero.range}
\usage{
zero.range(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{vector to check}

\item{tol}{tolerance}
}
\value{
logical
}
\description{
Determine if range of vector is FP 0 (from http://stackoverflow.com/questions/4752275/test-for-equality-among-all-elements-of-a-single-vector)
}
\examples{
x <- c(1, 2, 3, 4, 5, 6, 1)
y <- rep(2, times = 7)
zero.range(x) # FALSE
zero.range(y) # TRUE
}

