% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.spikes.R
\name{batch.spikes}
\alias{batch.spikes}
\title{batch.spikes}
\usage{
batch.spikes(rawdata, std.factor = 1, sigma = NA)
}
\arguments{
\item{rawdata}{ephys data to work on}

\item{std.factor}{depending on the quality of the data. A value of 1 is a good value for data with a really good signal to noise ratio.}

\item{sigma}{not yet implemented}
}
\value{
processed spike data
}
\description{
\code{batch.spikes} a first try on copying the functionality of the matlab toolbox batch_spikes.m function
more documentation might come in the future
steps: »unpack« the rawdata into filenames and data. find spikes in the data streams pack the data into one array
}

