% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.H1signal.R
\name{get.H1signal}
\alias{get.H1signal}
\title{get.H1signal}
\usage{
get.H1signal(data)
}
\arguments{
\item{data}{data from one recording file (2 channels)}
}
\value{
the data in a data frame
}
\description{
\code{get.H1signal} a copy of the matlab function with the same name. Since the data are not packed in a container together with metadata it is a simplified version.
}

