% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.data.R
\name{plot.data}
\alias{plot.data}
\title{plot.data}
\usage{
\method{plot}{data}(spikes, filelist, stimlist, outputdir)
}
\arguments{
\item{spikes}{data frame of spike data}

\item{filelist}{list of files to plot}

\item{stimlist}{currently not used}

\item{outputdir}{the output directory where the plots will be saved}
}
\value{
nothing yet
}
\description{
plot.data
}

