% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.isipsth.R
\name{plot.isipsth}
\alias{plot.isipsth}
\title{plot.isipsth}
\usage{
\method{plot}{isipsth}(spikes, binwidth, samplerate)
}
\arguments{
\item{spikes}{a spike data frame}

\item{binwidth}{}

\item{samplerate}{}
}
\value{
nothing yet
}
\description{
plot.isipsth
}

