% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.png.overview.R
\name{plot.png.overview}
\alias{plot.png.overview}
\title{plot.png.overview}
\usage{
\method{plot}{png.overview}(wd = NA, maxchunk = 100)
}
\arguments{
\item{wd}{character working directory where the data is. It is required to use the folder containing the data, meta, and png subfolders as input}

\item{maxchunk}{numeric add a maximum number of files to be processed at once}
}
\description{
\code{plot.png.overview} A function to create png overviews of ephys log files
}

