% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spikes.R
\name{plot.spikes}
\alias{plot.spikes}
\title{plot.spikes}
\usage{
\method{plot}{spikes}(spikes, filelist, stimlist, outputdir)
}
\arguments{
\item{spikes}{spike data}

\item{filelist}{list of files to plot}

\item{stimlist}{list of used stimuli}

\item{outputdir}{plot output direction}
}
\value{
nothing
}
\description{
plot spike data
}

