% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.spikes.R
\name{prune.spikes}
\alias{prune.spikes}
\title{prune.spikes}
\usage{
prune.spikes(spikes, min.isi)
}
\arguments{
\item{spikes}{spike matrix to be pruned}

\item{min.isi}{minimum inter stimulus interval in milliseconds}
}
\value{
a matrix containing a pruned spike
}
\description{
This Function removes spikes that fall within the minimum inter stimulus interval of the previous spike. It is safe to assume that a spike within 3 ms can be filtered out.
}
\details{
This file is part of the rephys package.

Copyright(c) Martin Rüßler

This file may be licensed under the terms of of the
GNU General Public License Version 3 (the ``GPL''),
or (at your option) any later version.

Software distributed under the License is distributed
on an ``AS IS'' basis, WITHOUT WARRANTY OF ANY KIND, either
express or implied. See the GPL for the specific language
governing rights and limitations.

You should have received a copy of the GPL along with this
program. If not, go to http://www.gnu.org/licenses/gpl.html
or write to the Free Software Foundation, Inc.,
51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.

The development of this software was supported by the
Excellence Cluster EXC 277 Cognitive Interaction Technology.
The Excellence Cluster EXC 277 is a grant of the Deutsche
Forschungsgemeinschaft (DFG) in the context of the German
Excellence Initiative.
}

