% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.metadata.R
\name{read.metadata}
\alias{read.metadata}
\title{read.metadata}
\usage{
read.metadata(files, datafolder, metafolder)
}
\arguments{
\item{files}{the files metadata should be loaded for}

\item{datafolder}{where the data files are located}

\item{metafolder}{where the metadata files are located}
}
\value{
returns a list with metadata for all files from the data folder
}
\description{
a function to read ephys metadata files.
}

